+++ src/NEventStore.Persistence.Sql.Tests/PersistenceTests.Transactions.cs
@@ -18,23 +18,38 @@ namespace NEventStore.Persistence.AcceptanceTests
     using System.Transactions;
     using System.Threading;
     using System.Globalization;
+    using System.Diagnostics;
 #endif
 #if XUNIT
     using Xunit;
     using Xunit.Should;
 #endif
 
+    public enum TransactionScopeConcern
+    {
+        NoTransaction = 0,
+        SuppressAmbientTransaction = 1,
+        EnlistInAmbientTransaction = 2
+    }
+
+    /// <summary>
+    /// This testing concern simulated the TrsanctionSuppression and/or re-enlisting in Ambient transaction behavior
+    /// </summary>
     public abstract class TransactionConcern : PersistenceEngineConcern
     {
-        protected void Reinitialize(bool includeInAmbientTransation)
+        protected void Reinitialize(TransactionScopeConcern enlistInAmbientTransation)
         {
-            if (includeInAmbientTransation)
+            switch (enlistInAmbientTransation)
             {
-                Fixture.ScopeOption = TransactionScopeOption.Required;
-            }
-            else
-            {
-                Fixture.ScopeOption = TransactionScopeOption.Suppress;
+                case TransactionScopeConcern.NoTransaction:
+                    Fixture.ScopeOption = null;
+                    break;
+                case TransactionScopeConcern.SuppressAmbientTransaction:
+                    Fixture.ScopeOption = TransactionScopeOption.Suppress;
+                    break;
+                case TransactionScopeConcern.EnlistInAmbientTransaction:
+                    Fixture.ScopeOption = TransactionScopeOption.Required;
+                    break;
             }
             Fixture.Initialize(ConfiguredPageSizeForTesting);
         }
@@ -46,19 +61,19 @@ namespace NEventStore.Persistence.AcceptanceTests
         protected const int Loop = 2;
         protected const int StreamsPerTransaction = 20;
         protected readonly IsolationLevel _transationIsolationLevel;
-        protected readonly bool _enlistInAmbientTransaction;
+        protected readonly TransactionScopeConcern _enlistInAmbientTransaction;
         protected Exception _thrown;
-        protected readonly bool _completeTRansaction;
+        protected readonly bool _completeTransaction;
 
-        public MultipleConnectionsWithMultipleTransactionScopes(
-            bool enlistInAmbientTransaction,
+        protected MultipleConnectionsWithMultipleTransactionScopes(
+            TransactionScopeConcern enlistInAmbientTransaction,
             IsolationLevel transationIsolationLevel,
-            bool completeTRansaction
+            bool completeTransaction
             )
         {
             _transationIsolationLevel = transationIsolationLevel;
             _enlistInAmbientTransaction = enlistInAmbientTransaction;
-            _completeTRansaction = completeTRansaction;
+            _completeTransaction = completeTransaction;
             Reinitialize(enlistInAmbientTransaction);
         }
 
@@ -88,7 +103,7 @@ namespace NEventStore.Persistence.AcceptanceTests
                             stream.CommitChanges(Guid.NewGuid());
                         }
                     }
-                    if (_completeTRansaction)
+                    if (_completeTransaction)
                     {
                         scope.Complete();
                     }
@@ -102,73 +117,27 @@ namespace NEventStore.Persistence.AcceptanceTests
     [TestClass]
 #endif
 #if NUNIT
-    [TestFixture(false, IsolationLevel.Serializable)]
-    [TestFixture(false, IsolationLevel.ReadCommitted)]
-    [TestFixture(true, IsolationLevel.Serializable)] // this will always fail! Serializable prevents multiple transation to insert simultaneously
-    [TestFixture(true, IsolationLevel.ReadCommitted)]
+    [TestFixture(TransactionScopeConcern.NoTransaction, IsolationLevel.Serializable)] // this will always fail! Serializable prevents multiple transation to perform insert queries simultaneously
+    [TestFixture(TransactionScopeConcern.NoTransaction, IsolationLevel.ReadCommitted)]
+    [TestFixture(TransactionScopeConcern.SuppressAmbientTransaction, IsolationLevel.Serializable)]
+    [TestFixture(TransactionScopeConcern.SuppressAmbientTransaction, IsolationLevel.ReadCommitted)]
+    [TestFixture(TransactionScopeConcern.EnlistInAmbientTransaction, IsolationLevel.Serializable)] // this will always fail! Serializable prevents multiple transation to perform insert queries simultaneously
+    [TestFixture(TransactionScopeConcern.EnlistInAmbientTransaction, IsolationLevel.ReadCommitted)]
 #endif
     public class Multiple_Completing_TransactionScopes_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is :
         MultipleConnectionsWithMultipleTransactionScopes
     {
         public Multiple_Completing_TransactionScopes_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is(
-            bool enlistInAmbientTransaction,
+            TransactionScopeConcern enlistInAmbientTransaction,
             IsolationLevel transationIsolationLevel
             ) : base(enlistInAmbientTransaction, transationIsolationLevel, true)
         { }
 
         [Fact]
-        public void should_throw_an_Exception_only_if_enlist_in_ambient_transaction_and_IsolationLevel_is_Serializable()
+        public void should_throw_an_Exception_only_if_no_transaction_or_enlist_in_ambient_transaction_and_IsolationLevel_is_Serializable()
         {
-            if (_enlistInAmbientTransaction && _transationIsolationLevel == IsolationLevel.Serializable)
-            {
-                _thrown.Should().BeOfType<AggregateException>();
-                _thrown.InnerException.Should().BeOfType<StorageException>();
-                // two serializable transactions on the same connection can result in deadlocks.
-                _thrown.InnerException.Message.Should().Contain("deadlock");
-            }
-            else
-            {
-                _thrown.Should().BeNull();
-            }
-        }
-
-        [Fact]
-        public void Should_have_no_commits_at_all()
-        {
-            if (_enlistInAmbientTransaction && _transationIsolationLevel == IsolationLevel.Serializable)
-            {
-                // nothing should be tested we never reach this point due to exception
-            }
-            else
-            {
-                _commits = Persistence.GetFrom().ToArray();
-                _commits.Length.Should().Be(Loop * StreamsPerTransaction);
-            }
-        }
-    }
-
-#if MSTEST
-    [TestClass]
-#endif
-#if NUNIT
-    [TestFixture(false, IsolationLevel.Serializable)]
-    [TestFixture(false, IsolationLevel.ReadCommitted)]
-    [TestFixture(true, IsolationLevel.Serializable)] // this will always fail! Serializable prevents multiple transation to insert simultaneously
-    [TestFixture(true, IsolationLevel.ReadCommitted)]
-#endif
-    public class Multiple_Failing_TransactionScopes_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is :
-        MultipleConnectionsWithMultipleTransactionScopes
-    {
-        public Multiple_Failing_TransactionScopes_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is(
-            bool enlistInAmbientTransaction,
-            IsolationLevel transationIsolationLevel
-            ) : base(enlistInAmbientTransaction, transationIsolationLevel, completeTRansaction: false)
-        { }
-
-        [Fact]
-        public void should_throw_an_Exception_only_if_enlist_in_ambient_transaction_and_IsolationLevel_is_Serializable()
-        {
-            if (_enlistInAmbientTransaction && _transationIsolationLevel == IsolationLevel.Serializable)
+            if (_enlistInAmbientTransaction != TransactionScopeConcern.SuppressAmbientTransaction
+                && _transationIsolationLevel == IsolationLevel.Serializable)
             {
                 _thrown.Should().BeOfType<AggregateException>();
                 _thrown.InnerException.Should().BeOfType<StorageException>();
@@ -184,21 +153,69 @@ namespace NEventStore.Persistence.AcceptanceTests
         [Fact]
         public void Should_have_expected_number_of_commits()
         {
-            if (_enlistInAmbientTransaction && _transationIsolationLevel == IsolationLevel.Serializable)
+            if (_enlistInAmbientTransaction != TransactionScopeConcern.SuppressAmbientTransaction
+                && _transationIsolationLevel == IsolationLevel.Serializable)
             {
-                // nothing should be tested we never reach this point due to exception
+                // unpredictable results: some transactions might succeed, other will deadlock
+                _commits = Persistence.GetFrom().ToArray();
+                _commits.Length.Should().BeGreaterThan(0);
             }
             else
             {
                 _commits = Persistence.GetFrom().ToArray();
-                if (!_enlistInAmbientTransaction)
-                {
-                    _commits.Length.Should().Be(Loop * StreamsPerTransaction);
-                }
-                else
-                {
-                    _commits.Length.Should().Be(0);
-                }
+                _commits.Length.Should().Be(Loop * StreamsPerTransaction);
+            }
+        }
+    }
+
+#if MSTEST
+[TestClass]
+#endif
+#if NUNIT
+    [TestFixture(TransactionScopeConcern.NoTransaction, IsolationLevel.Serializable)] // this will always fail! Serializable prevents multiple transation to perform insert queries simultaneously
+    [TestFixture(TransactionScopeConcern.NoTransaction, IsolationLevel.ReadCommitted)]
+    [TestFixture(TransactionScopeConcern.SuppressAmbientTransaction, IsolationLevel.Serializable)]
+    [TestFixture(TransactionScopeConcern.SuppressAmbientTransaction, IsolationLevel.ReadCommitted)]
+    [TestFixture(TransactionScopeConcern.EnlistInAmbientTransaction, IsolationLevel.Serializable)] // this will always fail! Serializable prevents multiple transation to perform insert queries simultaneously
+    [TestFixture(TransactionScopeConcern.EnlistInAmbientTransaction, IsolationLevel.ReadCommitted)]
+#endif
+    public class Multiple_Failing_TransactionScopes_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is :
+        MultipleConnectionsWithMultipleTransactionScopes
+    {
+        public Multiple_Failing_TransactionScopes_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is(
+            TransactionScopeConcern enlistInAmbientTransaction,
+            IsolationLevel transationIsolationLevel
+            ) : base(enlistInAmbientTransaction, transationIsolationLevel, completeTransaction: false)
+        { }
+
+        [Fact]
+        public void should_throw_an_Exception_only_if_enlist_in_ambient_transaction_and_IsolationLevel_is_Serializable()
+        {
+            if (_enlistInAmbientTransaction != TransactionScopeConcern.SuppressAmbientTransaction
+                && _transationIsolationLevel == IsolationLevel.Serializable)
+            {
+                _thrown.Should().BeOfType<AggregateException>();
+                _thrown.InnerException.Should().BeOfType<StorageException>();
+                // two serializable transactions on the same connection can result in deadlocks.
+                _thrown.InnerException.Message.Should().Contain("deadlock");
+            }
+            else
+            {
+                _thrown.Should().BeNull();
+            }
+        }
+
+        [Fact]
+        public void Should_have_expected_number_of_commits()
+        {
+            _commits = Persistence.GetFrom().ToArray();
+            if (_enlistInAmbientTransaction == TransactionScopeConcern.SuppressAmbientTransaction)
+            {
+                _commits.Length.Should().Be(Loop * StreamsPerTransaction);
+            }
+            else
+            {
+                _commits.Length.Should().Be(0);
             }
         }
     }
@@ -209,12 +226,12 @@ namespace NEventStore.Persistence.AcceptanceTests
         protected const int Loop = 2;
         protected const int StreamsPerTransaction = 20;
         protected readonly IsolationLevel _transationIsolationLevel;
-        protected readonly bool _enlistInAmbientTransaction;
+        protected readonly TransactionScopeConcern _enlistInAmbientTransaction;
         protected Exception _thrown;
         protected readonly bool _completeTransaction;
 
-        public MultipleConnectionsWithSingleTransactionScope(
-            bool enlistInAmbientTransaction,
+        protected MultipleConnectionsWithSingleTransactionScope(
+            TransactionScopeConcern enlistInAmbientTransaction,
             IsolationLevel transationIsolationLevel,
             bool completeTransaction
             )
@@ -229,30 +246,33 @@ namespace NEventStore.Persistence.AcceptanceTests
         {
             _thrown = Catch.Exception(() =>
             {
+                var eventStore = new OptimisticEventStore(Persistence, null);
+
                 using (var scope = new TransactionScope(TransactionScopeOption.Required,
                         new TransactionOptions { IsolationLevel = _transationIsolationLevel }
 #if NET451 || NETSTANDARD2_0
-                        , TransactionScopeAsyncFlowOption.Enabled
+                    , TransactionScopeAsyncFlowOption.Enabled
 #endif
-                    ))
+                ))
                 {
                     var res = Parallel.For(0, Loop, i =>
                     {
-                        var eventStore = new OptimisticEventStore(Persistence, null);
-
                         int j;
                         for (j = 0; j < StreamsPerTransaction; j++)
                         {
-                            using (var stream = eventStore.OpenStream(i.ToString() + "-" + j.ToString()))
+                            var streamId = i.ToString() + "-" + j.ToString();
+                            using (var stream = eventStore.OpenStream(streamId))
                             {
                                 for (int k = 0; k < 10; k++)
                                 {
                                     stream.Add(new EventMessage { Body = "body" + k });
                                 }
+                                Debug.WriteLine("Committing Stream: " + streamId);
                                 stream.CommitChanges(Guid.NewGuid());
                             }
                         }
                     });
+                    Debug.WriteLine("Completing transaction");
                     if (_completeTransaction)
                     {
                         scope.Complete();
@@ -263,17 +283,17 @@ namespace NEventStore.Persistence.AcceptanceTests
     }
 
 #if MSTEST
-    [TestClass]
+[TestClass]
 #endif
 #if NUNIT
-    [TestFixture(false, IsolationLevel.Serializable)]
-    [TestFixture(false, IsolationLevel.ReadCommitted)]
+    [TestFixture(TransactionScopeConcern.SuppressAmbientTransaction, IsolationLevel.Serializable)]
+    [TestFixture(TransactionScopeConcern.SuppressAmbientTransaction, IsolationLevel.ReadCommitted)]
 #endif
     public class Single_Completing_TransactionScope_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is
         : MultipleConnectionsWithSingleTransactionScope
     {
         public Single_Completing_TransactionScope_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is(
-            bool enlistInAmbientTransaction,
+            TransactionScopeConcern enlistInAmbientTransaction,
             IsolationLevel transationIsolationLevel
             ) : base(enlistInAmbientTransaction, transationIsolationLevel, completeTransaction: true)
         { }
@@ -293,17 +313,49 @@ namespace NEventStore.Persistence.AcceptanceTests
     }
 
 #if MSTEST
-    [TestClass]
+[TestClass]
 #endif
 #if NUNIT
-    [TestFixture(true, IsolationLevel.Serializable)] // unsupported: This platform does not support distributed transactions
-    [TestFixture(true, IsolationLevel.ReadCommitted)] // unsupported: This platform does not support distributed transactions
+    [TestFixture(TransactionScopeConcern.SuppressAmbientTransaction, IsolationLevel.Serializable)]
+    [TestFixture(TransactionScopeConcern.SuppressAmbientTransaction, IsolationLevel.ReadCommitted)]
 #endif
-    public class Unsupported_Single_Completing_TransactionScope_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is
+    public class Single_Failing_TransactionScope_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is
         : MultipleConnectionsWithSingleTransactionScope
     {
+        public Single_Failing_TransactionScope_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is(
+            TransactionScopeConcern enlistInAmbientTransaction,
+            IsolationLevel transationIsolationLevel
+            ) : base(enlistInAmbientTransaction, transationIsolationLevel, completeTransaction: false)
+        { }
+
+        [Fact]
+        public void should_not_throw_an_Exception()
+        {
+            _thrown.Should().BeNull();
+        }
+
+        [Fact]
+        public void Should_have_expected_number_of_commits()
+        {
+            _commits = Persistence.GetFrom().ToArray();
+            _commits.Length.Should().Be(Loop * StreamsPerTransaction);
+        }
+    }
+
+#if MSTEST
+[TestClass]
+#endif
+#if NUNIT
+    [TestFixture(TransactionScopeConcern.NoTransaction, IsolationLevel.Serializable)]
+    [TestFixture(TransactionScopeConcern.NoTransaction, IsolationLevel.ReadCommitted)]
+    [TestFixture(TransactionScopeConcern.EnlistInAmbientTransaction, IsolationLevel.Serializable)] // unsupported: This platform does not support distributed transactions
+    [TestFixture(TransactionScopeConcern.EnlistInAmbientTransaction, IsolationLevel.ReadCommitted)] // unsupported: This platform does not support distributed transactions
+#endif
+    public class Unsupported_Single_Completing_TransactionScope_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is
+           : MultipleConnectionsWithSingleTransactionScope
+    {
         public Unsupported_Single_Completing_TransactionScope_When_EnlistInAmbientTransaction_is_and_IsolationLevel_is(
-            bool enlistInAmbientTransaction,
+            TransactionScopeConcern enlistInAmbientTransaction,
             IsolationLevel transationIsolationLevel
             ) : base(enlistInAmbientTransaction, transationIsolationLevel, completeTransaction: true)
         { }
@@ -312,11 +364,25 @@ namespace NEventStore.Persistence.AcceptanceTests
         public void should_throw_an_StorageUnavailableException()
         {
             _thrown.Should().BeOfType<AggregateException>();
-            _thrown.InnerException.Should().BeOfType<StorageUnavailableException>();
-            _thrown.InnerException.Message.Should().Contain("This platform does not support distributed transactions");
+            AggregateException aex = _thrown as AggregateException;
+            aex.InnerExceptions
+                .Any(e => e.GetType().IsAssignableFrom(typeof(StorageUnavailableException)))
+                .Should().BeTrue();
+
+            //var storageExceptions = aex.InnerExceptions
+            //    .Where(e => e.GetType().IsAssignableFrom(typeof(StorageUnavailableException)))
+            //    .Select(e => e.Message);
+            //storageExceptions.Should()
+            //    .Match(c =>
+            //        c.Contains("This platform does not support distributed transactions.")
+            //        || c.Contains("The Promote method returned an invalid value for the distributed transaction.")
+            //    );
+
+            //.Contain("This platform does not support distributed transactions.");
+            // the following error means the transaction is being promoted to a distributed one, and still not supported
+            //    "The Promote method returned an invalid value for the distributed transaction.");
         }
     }
-
 }
 
 #pragma warning restore S101 // Types should be named in PascalCase